Rework runtime calculation
