+++ scripts/tcl_commands/cts.tcl
@@ -95,7 +95,7 @@ proc run_cts {args} {
 		check_cts_clock_nets
 		set ::env(cts_report_file_tag) $report_tag_holder
 		TIMER::timer_stop
-		exec echo "[TIMER::get_runtime]" >> [index_file $::env(cts_log_file_tag)_runtime.txt 0]
+		exec echo "[TIMER::get_runtime]" | python3 $::env(SCRIPTS_DIR)/write_runtime.py "cts"
 
 		set_def $::env(SAVE_DEF)
 		set ::env(CURRENT_SDC) $::env(SAVE_SDC)
